Move xpath grammar to xpath folder
